+++ User/app_config.h
@@ -6,12 +6,12 @@
 #define __APP_CONFIG_H
 
 #define APP_NAME              "Dev-Beta"
-#define APP_VERSION           "v0.4.2"
+#define APP_VERSION           "v0.4.3"
 #define APP_UI_NAME           "Astra"
 #define APP_TARGET_NAME       "STM32F103C8T6"
 #define APP_LCD_NAME          "ST7735S_128x160"
 #define APP_LCD_BOOT_NAME     "ST7735S 128x160"
-#define APP_USART_DEBUG_PORT  "COM9"
+#define APP_USART_DEBUG_PORT  "COM4"
 #define APP_USART_BAUDRATE    115200U
 
 #endif